Add support for multiple run and final reporters
This change adds support for final reporters, and raises the number of
run reporters to more than one. It also fixes a logic bug where the run
reporter was run at the end of the job instead of the end of the run.

The user can now add up several run and final reporters:
```
        "RunReporters": [
            {
                "Name": "TargetSuccess",
                "Parameters": {
                    "SuccessExpression": ">80%"
                }
            },
            {
                "Name": "noop"
            }
        ],
        "FinalReporters": [
            {
                "Name": "TargetSuccess",
                "Parameters": {
                    "AverageSuccessExpression": ">80%"
                }
            }
        ]
```

Signed-off-by: Andrea Barberio <barberio@fb.com>