Add initial support for run id in Status
This patch adds support for building a status report that takes into
consideration the run id. The events API still does not know anything
about the run id, so the support is not complete, but as soon as it
is added to the query interface, it would be just a matter of adding
a new query parameter.