Fix IsZero() check for go < 1.13
From the documentation:

To compare two Values, compare the results of the Interface method.
Using == on two Values does not compare the underlying values they
represent.

Also, QueryFields might not be comparable objects, so resort to
doing a check based on the type. We should be able to get rid
of this with go1.13 and value.IsZero()