Persist test descriptors to the storage layer
Currently we don't store the fetched test step descriptors, which makes
it hard to debug, audit, and resume jobs. This is the first building
block to achieve the above.

Also simplified pkg/test.Param type to be a string instead of a struct
wrapping on a string. No custom json marshal/unmarshal needed anymore.

See #75 .

Tested with a sample job on a sample DB:
```
cmds/clients/contestcli-http $ go run . start < start-literal.json
...
```

Retrieved the test descriptor:
```
$ echo 'use contest; select teststeps from jobs; order by job_id desc limit 1' | mysql -u contest -p contest -P 3306 -h localhost --protocol tcp -N | jq
[
  [
    {
      "Name": "cmd",
      "Label": "some label",
      "Parameters": {
        "args": [
          "Title={{ Title .Name }}, ToUpper={{ ToUpper .Name }}"
        ],
        "executable": [
          "echo"
        ]
      }
    }
  ]
]
```

Signed-off-by: Andrea Barberio <barberio@fb.com>