Simplify API for Job information storage and retrieval
There is an additional layer between the storage interface
and the interface the job manager uses to access job information.
This additional layer matches the events API, but it's not really
necessary for job information. The additional layer for events
API is necessary to introduce validation of some parameters
(e.g. the query) which should not be the plugin's responsibility.
But the JobStorage API does not require any further logic, so the
additional abstraction can be removed.