Expose events that have no target associated
This patch exposes test step events that have no targets associated, for
example events conveying plugin errors (as opposed to target failures).
These events are visible in the status report.
Note that previously events associated to a target were exposed in the
test step status. Now these events are exposed in the target status (a
sub-field of the test step status), while events that have no target are
exposd in the test step status.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>