[test_runner] Added support for running two pipelines back to back
This patch adds support for implementing a cleanup pipeline, which runs
cleanup steps, after the main test pipeline. One of the requirements of
ConTest has always been that it should support a "cleanup" phase, and
this adds support for it.

The cleanup pipeline is described in the same was test steps have been
described so far: and additional "cleanup" section in the test descriptor
is deserialized to implement a pipeline which runs the required cleanup
steps.