feat(targetmanager)!: EventEmitter for TargetManager
Initial problem:
It is required to display information about why
TargetManager acquired the targets it acquired, therefore
it is required to emit events right from a TargetManager

Implementation:
To do not add an extra entity for such case it is reached
a temporary agreement to use testevents with empty
TestStepLavel for TargetManager events. However to be
able to query such events it was required to do modify
testevent.Query.

BREAKING CHANGE:
Interface `TargetManager` was changed and it was moved to
`github.com/facebookincubator/contest/pkg/targetmanager`.

Signed-off-by: Dmitrii Okunev <xaionaro@fb.com>