Replace flag with pflag
We are not size-constrained for the client, so I am replacing the
standard `flag` with the more flexible `pflag`. Benefits:
* support GNU style short and long, (e.g. `-Y` and `--yaml` are now
  the same), which are more widely adopted
* positional arguments can occur before switches (e.g. now can run
  `contestcli-http start -Y` without generating an error)

This changes the default behaviour though, since long options require
a double-dash instead of a single one. Single-letter options (like `-r`)
are unchanged since they are treated as short options.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>