simple db based target locker
Summary of changes:
 - Add a minimal rdbms-based target locker implementation
 - Clarify target.Locker API docs


This adds a minimal target locker that uses a relational database as backend.

Using this locker allows sharing locks between multiple instances of Contest Server and makes locks survive restarts.

The unit/integration test suite for this new locker is an extended version of the inmemory test suite, some tests changed to confirm with the clarified locker API and some tests added to prevent bugs like #134.