[test_runner] Fix deadlock in routing code
Routing now consists of input routing and output routing.
If any of the two blocks fails, as we are seeing now with a bug
in ForEachTarget which serializes targets injection, the error should
be considered as fatal, and a termination signal should be propagated
to all the routing logic, the router should return and trigger a
cancellation of the whole test.

Signed-off-by: Marco Guerri <marco.guerri.dev@fastmail.com>