Fix race condition in job status emission
Before this patch there is a race condition where the job status event
is emitted before a report is saved. A client waiting on the status
might get zero job reports if the job status is fetched before the
report is saved to the database. This is a behaviour we have seen
internally.

Signed-off-by: Andrea Barberio <insomniac@slackware.it>