[RFC] Schema migration tool
Proposal for a schema migration tool, based on `golang-migrate/migrate`.

Each migration is representd with a Task structure, which dictates
which schema changes should be applied and defines in code how
data should be migrated from prevoius to new schema. So, the migration
process consists in two parts:

* Schema migration: this is done via golang-migrate/migrate library.
SQL for forward and backward migration is defined Task structures.
The schema is versioned based on the version number returned by
First/Next/Prev functions implemented by `Task` structures. Migrations
are applied in ascending order by `golang-migrate/migrate`, which keeps
track of the current version of the schema, and whether it was
applied cleanly in a dedicated table (`schema_migrations`).
* Data migration: after a schema has been migrated, data might need
to be migrated as well. This is done via also in code by `Task` structs
The same approach for tracking data migrations as schema migrations
will be implemented: there will be a `data_migrations` table which
keeps track of which schema version the data has been migrated to.
Schema migration and data migration versions need to be aligned,
even though it is not mandatory to define code to implement a data
migration.