[pkg/teststeps] Avoid busy loop in ForEachTarget
I've hit a bug where cancellation leads to a busy loop until all the
plugins have returned. This happens because once the cancellation is
requested, the for-select will continuously enter the cancellation case.
With this sleep the cancellation path is not busy anymore and the CPU
doesn't suffer.
This was particuarly bad with buggy plugins that do not return quickly
enough (or at all - which is a plugin bug).

Signed-off-by: Andrea Barberio <insomniac@slackware.it>