feat(jobmanager): Add support of "wait and exit".
Previously we had only "pause all the jobs and exit" signal for
a JobManager. But since "pause" is still not implemenented
neither in the core code nor in plugins, it is required to
introduce another way to graceful shutdown a ConTest instance.
And meanwhile it is proposed to use "wait for all the job
to complete and the exit".