[db][migration] Implement db migration during container initialization
With support for schema migration, we'll start tracking db schema
versions and migration scripts. This implies using v0 as the base
schema version and then apply all available migrations to bring
the schema up to date. The latest version will not be tracked directly.
This patch implements this mechanism in mysql container.