[db][migration] Improve extended descriptor migration
A significant limitation of the library we use for db migrations is
that a migration can run only within a transaction. This doesn't scale
for large migrations. This patch introduces an interface to run a
migration outside of a `*sql.Tx`. The actual tool which invokes the
migration does not support running the transaction within `*sql.DB` yet:
we either need to add support to the upstream library, or move to a
custom implementation.

In addition, this patch modifies extended descriptor migration to run
batch update queries. The batch and single row implementations were
tested against ~95k records and resulted in roughly the following
numbers:
* Batch update (shard size 50): 4m26s
* Single row update (shard size 50): 5m51s