[storage] Enforce the desired db version
We started tracking the db schema version. The code will have a
hard dependency on a specific schema version being available, so
we should fail any attempt to set a storage backend if its advertised
schema version does not match the requirements of core ConTest storage
pkg..

For all those instances for which schema versioning is not in place yet
(e.g. schema is v0, so the reference baseline), reported version will
be 0. This patch sets the minimum schema version to 0, so it will allow
any backend version.