[contest] Introduce extended_descriptor
The way descriptors are stored in the request table at the moment
requires that we go again through test fetchers every time
we want to re-use a previously persisted descriptor. This because
even if we have the steps description persisted as part of `teststeps`,
that doesn't include the test names. If the test description is inlined
in the job descriptor, the retrieval with literal plugin is a no-op.
However, in any other case where we need to fetch the description from
a backend, this might become a problem.