Introduce a buffer into target result channel
Buffer of 1 is needed so that the step reader does not block when submitting result back
to the target handler. Target handler may not yet be ready to receive the result,
i.e. reporting TargetIn event which may involve network I/O.

1 is also sufficient because each target is running through only one step at each time,
so there won't be more than one result.