TargetLocker.Refresh and .Unlock API changes, clean up tests
Change the semantics of TargetLocker.Refresh to require that owner holds
locks for the requested targets (having an expired lock for the same job
is also permitted).

This is necessary for job resumption: when resuming a paused job we need
to be sure that targets have not in the mean time been assigned to other
jobs and thus had their state interfered with.

Some drive-by improvements:

 * Added TargetLocker.Close() to clean up locker instances. Used in tests.
 * Parametrise lock lease length, expose as `-targetLockDuration`
   argument and increase the default to 10 minutes.
 * Unify testing for TargetLocker implementations: use the same suite of
   tests for both inmemory and dblocked implementations, resolved subtle
   differences exposed by this.
 * Passing empty target list to the locker API is now an error (used to
   be allowed by dblocked but not inmemory).
 * Trying to unlock a target that was not locked by the specified job is
   an now an error.
 * Wire clock.Clock through job manager and runner API, for use in tests.
   Actual use is a TODO (turned out to be non-trivial)
 * Speed up job manager shutdown (speeds up tests)