Add end to end tests, with real server and client
Client and server binary are refactored to be reusable within the same process,
tests/e2e/e2e_test starts server on a random port, uses client to submit jobs and check status.
This is as close to using standalone binaries as possible but still provides coverage.

Two E2E tests are implemented:
 * A simple test that runs a single job in one go and checks its status.
 * A pause/resume test that starts a job and then stops server multiple
   times to force pause and resume, then verifies that all teh steps
   have been executed once and in the correct order.

Signed-off-by: Deomid "rojer" Ryabkov <rojer9@fb.com>