Add support for intra-step pause
If asked to pause, a step can return ErrPaused with a piece of state
that will be passed to a new instance when the job is resumed.
The state is opaque but at a minimum should contain targets that were in
flight at the time, they will not be reinjected and runner will expect
results for them when the job is resumed.

Handling pause is not a requirement, a step can completely ignore
the pause signal and pausing will still work between steps.