dblocker: Fix rare race in TryLock
Two attempts to (Try)Lock a set of devices can race.
They get rejected by the database as expected with "duplicate insert...".
This works fine, but is a problem in TryLock where we explitily want to do best effort locking.

To address this, TryLock will now use INSERT IGNORE, and then read the locks back to see what got through.