Fail zombie jobs on startup
Jobs in the "started" state and that belong to the server instance on
startup are zombies most likely resulting from improper server shutdown.
These are forced into "failed" state and a log entry is emitted for them
to provide some sort of closure and make it less confusing to the users.