Support MariaDB
This adds support for MariaDB as an alternative to MySQL.

MariaDB must be selected explicitly by pointing the tooling to `docker-compose.mariadb.yml`, or running the test script with `run-test.sh -d mariadb`.
There are no plans to change the default.

Primary motivation is that MySQL has no official images for non-x86 architectures in Docker Hub. (Full support for this in the next PR)